+++ back/core/src/main/java/ports/repositories/DevisRepoPT.java
@@ -10,8 +10,8 @@ public interface DevisRepoPT {
 
     DevisDN save(DataProviderManager dpm, DevisDN devis) throws PersistenceException;
 
-    int countDevisOfMonth(DataProviderManager dpm,Date date);
+    Long countDevisOfMonth(DataProviderManager dpm,Date date);
 
-    int existsNumeroDevis(DataProviderManager dpm,String numeroDevis);
+    Long existsNumeroDevis(DataProviderManager dpm,String numeroDevis);
 
 }
